refactor: separate repository interfaces
